feature: Create map context to avoid prop drilling && some code cleanup
